requirements.txt
acres==0.5.0
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
ci-info==0.4.0
click==8.3.1
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
etelemetry==0.3.1
filelock==3.25.2
frontend==0.0.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jiter==0.13.0
looseversion==1.3.0
lxml==6.0.2
networkx==3.6.1
nibabel==5.4.2
nipype==1.11.0
numpy==2.4.3
openai==2.29.0
packaging==26.0
pageindex==0.2.8
pandas==3.0.1
pathlib==1.0.1
pillow==12.1.1
prov==2.1.1
puremagic==2.1.0
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
PyMuPDF==1.27.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyxnat==1.6.4
rdflib==7.6.0
requests==2.32.5
scipy==1.17.1
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
traits==7.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
